safari-534.53-branch/LayoutTests/fast/inspector-support/uncaught-dom1-exception.html

<html>
<head>
<script>

function onload()
{
    if (window.layoutTestController)
        layoutTestController.dumpAsText();

    // Following code generates an exception.
    var p = document.createTextNode();
    p.textContent = "foo";
    p.splitText(10000);
}

</script>
</head>

<body onload="onload()">
<p>
This tests that we are getting the correct message for DOM Exception 1: INDEX_SIZE_ERR.
</p>

</body>
</html>
