LayoutTests/platform/chromium/fast/inspector-support/uncaught-dom1-exception-expected.txt

CONSOLE MESSAGE: line 13: Uncaught Error: INDEX_SIZE_ERR: DOM Exception 1
This tests that we are getting the correct message for DOM Exception 1: INDEX_SIZE_ERR.
